Optimize file handling and update version
- Enhanced DDS file handling in MenuArtConverter.cs by adding a check to bypass conversion for files already in DDS format.
- Improved file conversion logic in PictoConverter.cs to include a DDS check, mirroring the optimization in MenuArtConverter.cs, and adjusted control flow for .xtx file deletion.
- Adjusted the data type of `TrackId` from `int` to `long` in KaraokeTape.cs because some jsons add .0 to numbers.
-  Version bump

+++ JustDanceEditor.Converter/UbiArt/Tapes/KaraokeTape.cs
@@ -15,7 +15,7 @@ public class KaraokeClip
 {
     public string __class { get; set; }
     public long Id { get; set; }
-    public int TrackId { get; set; }
+    public long TrackId { get; set; }
     public int IsActive { get; set; }
     public int StartTime { get; set; }
     public int Duration { get; set; }
